+++ fluxpype/__init__.py
@@ -1,1 +1,8 @@
-from fluxpype.runners import config_runner_timeoutsafe as config_runner
+# fluxpype/__init__.py
+
+import os
+from pathlib import Path
+from warnings import warn
+
+# from . import config_runner
+from .config_runner import run
