--- fluxpype/config/config.ini
@@ -1,79 +1,0 @@
-[DEFAULT]
-config_name = velocityplots  # Name of the configuration to be used by default
-
-jsoc_email = chris.gilly@colorado.edu  # Email to use for JSOC data queries
-
-# Base Paths
-base_dir = ~/vscode/fluxons  # Define the base directory here
-
-# Paths (use base_dir for dynamic reference)
-rc_path = ~/.zshrc  # Location of .zshrc file for shell configuration
-run_script = %(base_dir)s/fluxpype/fluxpype/science/magnetogram2wind.pdl  # Path to the run script for the pipeline
-python_dir = %(base_dir)s/fluxpype/.venv/bin/python3  # Path to Python executable
-pl_prefix = ~/Library/perl5  # Perl scripts location prefix
-fl_prefix = ~/Library/flux  # Location of fluxon-MHD code
-fl_mhdlib = %(base_dir)s  # Location of fluxon-MHD code
-data_dir = %(base_dir)s/fluxpype/fluxpype/data  # Location where fluxon-data is stored
-working_dir = %(base_dir)s/fluxpype
-
-#############################################
-# Default Batch parameters
-batch_name = default_batch  # Name for the default batch to be processed
-rotations  = [2110]  # List of solar rotations to process
-fluxon_count = [1000]  # Number of fluxons for each rotation
-
-# Parameters for MHD modeling
-mag_reduce = 2  # Factor by which to reduce magnetic field data
-inner_fov = 2.5  # Inner field of view in solar radii
-outer_fov = 6.0  # Outer field of view in solar radii
-
-# Relaxation parameters
-do_steps = 400  # Number of steps to perform in each relaxation cycle
-relax_threshold = 1.5  # Threshold for stopping relaxation
-max_cycles = 10  # Maximum number of relaxation cycles
-
-# Default run flags
-verbose = 0  # Enable verbose logging (0 for off, 1 for on)
-do_download = 0  # Download new data (0 for off, 1 for on)
-do_wind_calc = 1  # Calculate solar wind (0 for off, 1 for on)
-plot_only = 0  # Only plot the data without computation (0 for off, 1 for on)
-recompute = 0  # Force re-computation of existing data (0 for off, 1 for on)
-adapts = [0] #use the ADAPT maps instead of HMI (0 for off, 1 for on)
-adapt_select = 2 #select which ADAPT map to use (0-12)
-
-#############################################
-## Test configuration sections
-
-[TestConfig]
-batch_name = TestConfig_batch
-rotations  = [2225]
-fluxon_count = [100]
-adapts = [0, 1]
-mag_reduce = 4
-
-
-[Multi_Test_Config]
-batch_name = Test_config_batch
-rotations  = [2101, 2122]
-fluxon_count = [1000, 1500]
-
-#############################################
-## Real configuration sections
-
-[Real_Run]
-batch_name = For_real_this_time
-rotations  = [2273]
-mag_reduce = 4 # Factor by which to reduce magnetic field data
-fluxon_count = [200]
-adapts = [0, 1]
-adapt_select = 2
-
-[velocityplots]
-batch_name = Newtest
-rotations  = [2168]  #, 2250, 5) #(2190, 2230, 1) # (start, stop, step) when parens are used
-fluxon_count = [100]
-# Wind Modeling parameters
-flow_method = [wsa] #, wsa, ghosts, parker, cranmer] #, ghosts, wsa, parker] # can be [tempest, ghosts, wsa, parker, cranmer]
-concurrency = 1  # Number of concurrent processes to run
-recompute_wind = 0 # Force re-computation of existing wind data (0 for off, 1 for on)
-wind_only = 0   # Skip fluxon calculation (0 for off, 1 for on)
