--- fluxpype/runners/_Inline/config-darwin-2level-5.032000
@@ -1,20 +1,0 @@
-version : 0.86
-languages : %
-    C : C
-    Foo : Foo
-    PDLPP : Pdlpp
-    Pdlpp : Pdlpp
-    foo : Foo
-    pdlpp : Pdlpp
-types : %
-    C : compiled
-    Foo : interpreted
-    Pdlpp : compiled
-modules : %
-    C : Inline::C
-    Foo : Inline::Foo
-    Pdlpp : Inline::Pdlpp
-suffixes : %
-    C : bundle
-    Foo : foo
-    Pdlpp : bundle
