+++ fluxpype/science/get_magnetograms.py
@@ -28,12 +28,12 @@ Gilly <gilly@swri.org> (and others!)
 
 
 import argparse
-from fluxpype.helpers.pipe_helper import (configurations, get_magnetogram_file, get_ADAPT_file)
+from pipe_helper import (configurations, get_magnetogram_file, get_ADAPT_file)
 configs = configurations()
 # Create the argument parser
 parser = argparse.ArgumentParser(description='This script downloads a magnetogram for a particular Carrington Rotation')
 parser.add_argument('--cr', type=int, default=2100, help='Carrington Rotation')
-parser.add_argument('--datdir', type=str, default='/Users/cgilbert/vscode/fluxons/fluxon-data', help='data directory')
+parser.add_argument('--datdir', type=str, default='/Users/cgilbert/vscode/fluxons/fluxpype/fluxpype/data', help='data directory')
 parser.add_argument('--reduce', type=int, default=5, help='factor by which the magnetogram is reduced')
 parser.add_argument('--do_download', type=int, default=0, help='download the files')
 parser.add_argument('--adapt', type=int, default=1, help='download the files')
