+++ fluxpype/science/get_pfss.py
@@ -36,7 +36,7 @@ import os
 import argparse
 import numpy as np
 from fluxpype.science.pfss_funcs import trace_lines, load_pfss, compute_pfss, load_and_condition_fits_file, get_fluxon_locations
-from fluxpype.helpers.pipe_helper import shorten_path, configurations
+from fluxpype.pipe_helper import shorten_path, configurations
 import astropy.constants as const
 import astropy.units as u
 import matplotlib.colors as mcolor
@@ -62,7 +62,7 @@ def get_pfss(configs=None):
 
     configs = configs or configurations()
 
-    from fluxpype.helpers.pipe_helper import update_magdir_paths
+    from fluxpype.pipe_helper import update_magdir_paths
     configs = update_magdir_paths(configs)
 
     # Extract arguments or use defaults from configs
@@ -179,7 +179,7 @@ def get_regular_pfss(configs=None):
 
     configs = configs or configurations()
 
-    from fluxpype.helpers.pipe_helper import update_magdir_paths
+    from fluxpype.pipe_helper import update_magdir_paths
     configs = update_magdir_paths(configs)
 
     # Extract arguments or use defaults from configs
@@ -326,7 +326,7 @@ def get_regular_pfss2(configs=None):
 
     configs = configs or configurations()
 
-    from fluxpype.helpers.pipe_helper import update_magdir_paths
+    from fluxpype.pipe_helper import update_magdir_paths
     configs = update_magdir_paths(configs)
 
     # Extract arguments or use defaults from configs
