+++ fluxpype/science/make_world_from_pfss.pm
@@ -16,13 +16,13 @@ use PDL::NiceSlice;
 use PDL::Options;
 use Flux;
 use PDL::IO::Misc;
-use make_world_sphere qw(make_world_sphere);
+# use make_world_sphere qw(make_world_sphere);
 use File::Path;
 use Time::HiRes qw(clock_gettime);
 use File::Basename qw(fileparse);
 use File::Basename;
-use pipe_helper qw(shorten_path);
-use plot_world qw(plot_world);
+# use pipe_helper qw(shorten_path);
+# use plot_world qw(plot_world);
 use Flux::World    qw(read_world);
 use Flux::World qw(str2world);
 use PDL::Graphics::Gnuplot qw(gpwin);
